+++ 2024/rust/src/main.rs
@@ -9,7 +9,7 @@ fn usage() -> ! {
     std::process::exit(1);
 }
 
-fn solve(part: u32, part1_fn: fn(&str) -> i64, part2_fn: fn(&str) -> i64, input: &str) {
+fn solve(part: u32, part1_fn: fn(&str) -> usize, part2_fn: fn(&str) -> usize, input: &str) {
     match part {
         1 => {
             let answer1 = part1_fn(input);
